Move manager/internal/config/trafficManager.go to manager/state/mode.go
Mode is needed when extending the traffic-manager.
