+++ cmd/traffic/cmd/manager/manager.go
@@ -27,9 +27,9 @@ import (
 	"github.com/datawire/dlib/dlog"
 	"github.com/datawire/k8sapi/pkg/k8sapi"
 	rpc "github.com/telepresenceio/telepresence/rpc/v2/manager"
-	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/internal/config"
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/internal/mutator"
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/managerutil"
+	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/state"
 	"github.com/telepresenceio/telepresence/v2/pkg/tracing"
 	"github.com/telepresenceio/telepresence/v2/pkg/version"
 )
@@ -112,7 +112,7 @@ func Main(ctx context.Context, _ ...string) error {
 func (m *Service) configMapEventHandler(eventType watch.EventType, obj runtime.Object) error {
 	if eventType == watch.Added || eventType == watch.Modified {
 		var (
-			tmConf    config.TrafficManager
+			tmConf    state.TrafficManager
 			yamlBytes = m.configWatcher.GetTrafficManagerConfigYaml()
 		)
