+++ cmd/traffic/cmd/manager/state/mode.go
@@ -1,4 +1,4 @@
-package config
+package state
 
 import (
 	"fmt"
