+++ cmd/traffic/cmd/manager/state/state.go
@@ -17,7 +17,6 @@ import (
 
 	"github.com/datawire/dlib/dlog"
 	rpc "github.com/telepresenceio/telepresence/rpc/v2/manager"
-	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/internal/config"
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/internal/watchable"
 	"github.com/telepresenceio/telepresence/v2/cmd/traffic/cmd/manager/managerutil"
 	"github.com/telepresenceio/telepresence/v2/pkg/log"
@@ -54,7 +53,7 @@ type State struct {
 	timedLogLevel   log.TimedLevel
 	llSubs          *loglevelSubscribers
 	cfgMapLocks     map[string]*sync.Mutex
-	config          config.TrafficManager
+	config          TrafficManager
 }
 
 func NewState(ctx context.Context) *State {
@@ -633,7 +632,7 @@ func (s *State) WaitForTempLogLevel(stream rpc.Manager_WatchLogLevelServer) erro
 	return s.llSubs.subscriberLoop(stream.Context(), stream)
 }
 
-func (s *State) SetConfig(c config.TrafficManager) {
+func (s *State) SetConfig(c TrafficManager) {
 	s.mu.Lock()
 	oldMode := s.config.Mode
 	s.config = c
